[REFACTOR][ABI] Update backtrace storage for append-friendly (#39)
This PR updates the default backtrace order convention to most recent
call first. Such storage format makes it easy to append the backtrace
and incrementally build up as we propagate the error up.

Importantly, the previous rationale of using traceback(most recent call
last) is to align with python style error handling. This is still
remains an important design consideration. As a result, when we render
the error or propagate the result, we will still call
TracebackMostRecentCallLast, so such change is not visible to the user,
but will enable more memory efficient incremental building of backtrace.

This PR also updates the naming convention accordingly. We also updated
TVMFFIErrorCreate to be more consistent with other parts.

Also fixes minor nits in quick start to simplify dependency
installation.

This is an ABI change. Given we are still in pre-release beta, we bump
beta version number.