[FFI][ABI] TVMFFIObject ABI to further align with torch intrusive_ptr (#57)
This PR updates the TVMFFIObject ABI to use
strong_ref_count/weak_ref_count
into u32 and which can be packed into the first u64 field in the object
header.

This change will open up further optimizations that combines weak/strong
into
a single u64 atomic that skips the need of extra atomic reading of weak
value
during deletion. The technique was recently in torch

We also reordered the counter to be first since most rc ops are frequent
and
it would be more cache friendly. We do give up the same type_index first
location
with TVMFFIAny, which is OK as they optimize for different goals.