[CYTHON] Enable ffi function creation from MLIR Packed function (#100)
MLIR Execution engine generates function that corresponds to the packed
signature. As of now it is hard to access the raw extern C function
pointer of SafeCall directly when we declare the signature in LLVM
dialect.

This helper enables us to create ffi.Function from safe call function
pointer that are exposed as MLIR packed function pointer calling
convention. It can help facilitate DSLs that leverages MLIR execution
engine to do JIT.

Note that in theory MLIR execution engine should be able to also support
some form of "extern C" feature that directly exposes the funtion
pointers of C-compatible functions with an attribute tag. So we keep
this feature in the python helper layer for now in case MLIR execution
engine supports it in the future.