fix: Add schema support for `ObjectRef::MemFn` (#113)
This PR adds specialization of `tvm::ffi::details::FunctionInfo<>` for
`ObjectRef` types, which fixes downstream compilation error.

For more context, in downstream TVM usecases, a member function of an
`ObjectRef` could be registered as a global function, which is not taken
into consideration in current project, resulting in compilation issues.