[CYTHON][REFACTOR] __tvm_ffi_object__ protocol (#115)
This PR does two things:
- We renamed `PyNativeObject.__tvm_ffi_object__` to
`_tvm_ffi_cached_object` This is because the dunder convention is
normally used for functions instead of attributes
- We then updated the original `__tvm_ffi_tensor__` protocol to
`__tvm_ffi_object__` so that the same protocol can be used for generic
objects.

Given the original attribute is hidden to user and we did not yet
introduce `__tvm_ffi_tensor__` in a formal release, the change should
have low impact. We will however freeze the protocol for formal release
ideally soon.