[CPP] Remove static line based object reg in favor of ObjectDef (#116)
This PR moves static inline registrations of objects in the object reg
macros.

Rationale: while the static inline based registration is convenient and
removes the need for explicit registration this can cause extra space
overhead for every dll that includes the header file even if the class
is not used.

Given that reflection::ObjectDef<T> already registers the type index and
static type objects are predefined that we can control, we explicitly
removes the static inline based reg trigger to we can save up possible
binary space and reduce init logic.

This does mean that when an Object is not registered through ObjectDef
the dynamic casting mechanism won't work. But we are OK since
expectation is most dynamic ObjectRef needs to do so for it to be
useful.