[TEST] Fix the tensor loading order in test (#128)
If we load a module within a local scope,
that object's deleter resides in the loaded library. if the module is
unloaded before the object is destroyed, the deleter may call an invalid
address. Keep the module loaded until all returned objects are deleted.

This PR fixes the testcase that have the related issues