[ABI] Refactor the TVMFFIEnvTensorAllocator to align with DLPack (#131)
This PR refactors the TVMFFIEnvTensorAllocator to align with the
finalized naming of DPack standard.

- TVMFFIEnvSetTensorAllocator ->
TVMFFIEnvSetDLPackManagedTensorAllocator
- TVMFFIEnvGetTensorAllocator ->
TVMFFIEnvGetDLPackManagedTensorAllocator

We also introduced TVMFFIEnvTensorAlloc to directly allocate a
ffi::Tensor from the given prototype which can be used by DSL compilers.

Note that the new TVMFFIEnvTensorAlloc makes metadata allocation in
libtvm_ffi so it won't suffer from the module unloading order problem.
We removed Tensor::FromDLPackAlloc in favor of Tensor::FromEnvAlloc that
makes use of TVMFFIEnvTensorAlloc.

Hopefully this step allows us to stablize the tensor alloc api to align
with dlpack before we freeze.